DOC: CLI updates for align, call, div; FDR + call
Makes sense to make FDR curve generation a part of calling, since
we can't really easily create these curves for other variant calling
methods (because how do we vary some sort of parameter like p)?

Eventually we'll need to reconcile the paper's description of
computing the minimum sufficient coverage from "m" and "p", but for
now (when we're working with arbitrary variant calls) we just let
the user set this coverage directly. Probs we can get call-naive to
output info about what value of p it settled on, and then the user
can optionally use that to inform the coverage here... idk