Fix minor get_quartet() bug, flake8 fixes
the quartet thing -- the parameter was named alnseg and i forgor
to update the body of the function to compensate, so it was still
referring to linearaln. however, linearaln was passed *as* alnseg, so
this really shouldn't have made a difference -- it would've started to
cause errors now, since linearaln will no longer be in global scope
(since it's now wrapped up in the scope of filter_osa_reads()).