Handle "empty" seqs more gracefully in filters
The OSA and PM filters, specifically. Closes #11.

division by zero was what killed the PM filter on the cluster. It's
weird that this didn't cause problems when I ran this as a standalone
script months ago...? Oh, that's it -- since i was only updated during
the enumerate() step at
https://github.com/fedarko/sheepgut/blob/ee74a550145c5e0c2fb56b8ed0484b7df6decc1b/main-workflow/filter-partially-mapped-reads.py#L127-L130
it just straight up wasn't getting updated at all when focused_seq
was empty -- in these cases i would just remain the same as previously
in the loop, so the update messages being printed out were incorrect.
Huh! Weird. Well, fortunately it's only a bug with the update messages
-- the actual output should be fine.