ENH: Flesh out skeleton for hotspot finding
Also, after experimenting with prodigal for a bit, it makes sense to
just pass in all of the genes in a single GFF file, and merge
hot-one and hot-all into a single command. That way the burden of
gene finding is completely moved to the user, at least for this
particular context.