ENH/DOC: Revise structure of hotspot/coldspot cmds
I think it makes sense to 1) move to just identifying hotspot
"features" and ignoring distinctions between genes, exons, etc.
for this use-case (since this is the easiest way to handle this
in a correct way), and 2) to rename the commands to both be clear
that these are only finding certain types of hotspots / coldspots.