MNT: allow degen nucleotides in reads: just ignore
Closes #12. Since all of the code defines coverage at a position in
pysamstats as A+C+G+T, degen nucleotides aligned to a position are
implicitly ignored -- and, thanks to the existing validation applied
to input FASTA files and BCF files, we know that there will not be
degen nucleotides within contig sequences or in mutations. So, we
should be good. updating paper also.