MNT/DOC: abstract VERBOSE descs in some places
in a few places, the parts of the code that do logging-only-if-verbose
essentially span the entire command's execution, so it makes sense
to just say "while running". but in some cases we only do this type
of logging in certain regions, e.g. in "align" we only do this during
alignment filtering (and not while minimap2 is running or whatevs)