ENH: matrix fill skeleton & params; fancystart bug
applying format (e.g. f"{p:,}") to values that might be None causes
ugly crashes -- ideally we'd still format these values if they are
passed (shouldn't be difficult to detect this, although it would
require finagling fancystart a bit), but for now omitting the special
numerical formatting resolves the problem.