DOC: Various wording things in tut intro;rerunning
(reran gfa-to-fasta and align, just to make sure that this had been
done with >= v0.2.0 of strainFlye)

notably, the align output is complete (because i kept my laptop logged
in the entire time it was running lol) -- but at the cost of making
the notebook a bit cumbersome due to all the minimap2 output. Might
be nice to remove some of that from the tutorials?