DOC: another linebreak in the flow chart
one more thing i forgot to mention in the previous commit -- using
"#" in the node name for "# mutations / Mb" caused the node label
to vanish, presumably because of
https://github.com/mermaid-js/mermaid/issues/5419. the previous commit
fixed this (changing "# mutations" to "Number of mutations"), but this
made the node label a bit too wide (not wide enough to trigger the
other bug documented in the previous commit, but still wide enough to
look a bit off). so anyway this commit adds another <br/> to make the
nodes all have roughly the same size.